% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_calibration_wavelets.R
\name{Logistic_Density}
\alias{Logistic_Density}
\title{Logistic Density}
\usage{
Logistic_Density(theta, tau)
}
\arguments{
\item{theta}{Numeric value of \eqn{\theta}.}

\item{tau}{Numeric value of \eqn{\tau}.}
}
\value{
A numeric value representing the result of the function \eqn{g(\theta; \tau)} for the specified inputs.
}
\description{
Computes the function:

\deqn{g(\theta; \tau) = \displaystyle \frac{\exp\left(-\frac{\theta}{\tau}\right)}{\tau \left(1 + \exp\left(-\frac{\theta}{\tau}\right)\right)^2}}
}
\keyword{internal}
