% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_calibration_wavelets.R
\name{Universal_Shrinkage}
\alias{Universal_Shrinkage}
\title{Universal Shrinkage}
\usage{
Universal_Shrinkage(d, lambda)
}
\arguments{
\item{d}{Numeric value of the empirical coefficient to be denoised}

\item{lambda}{Numeric value of \eqn{\lambda}}
}
\value{
A numeric value representing the result of the Universal shrinkage applied to the empirical coefficient \eqn{d}.
}
\description{
A short description...
}
\keyword{internal}
